Migrated to use a http service layer

+++ features/addfood/src/main/java/com/scottandrews/addfood/data/model/AddFoodModel.kt
@@ -3,5 +3,5 @@ package com.scottandrews.addfood.data.model
 data class AddFoodModel(
     val barcode: String,
     val protein: Double,
-    val kcal: Double,
+    val kcal: Long,
 )
